+++ heap_soft/heap_sort.c
@@ -1,4 +1,4 @@
-#include "heap_soft.h"
+#include "heap_sort.h"
 
 /**
  * @brief Troca os valores de dois inteiros.
@@ -56,6 +56,11 @@ void construir_heap(heap_t *heap) {
  *       e reorganiza o heap para os elementos restantes, repetindo até ordenar todo o vetor.
  */
 void heap_soft(heap_t *heap) {
+
+    if (heap->tamanho <= 1) {
+        return; // Se a heap já está vazia ou tem apenas um elemento, não há nada a fazer
+    }
+
     construir_heap(heap);
     for (int i = heap->tamanho - 1; i >= 1; i--) {
         trocar(&heap->dados[0], &heap->dados[i]);
